Rename hardware interface files

+++ spot_ros2_control/hardware/include/spot_ros2_control/spot_hardware_interface.hpp
@@ -95,7 +95,7 @@ class SpotHardware : public hardware_interface::SystemInterface {
  private:
   // TODO(khughes): Hard coding this for now, but there should be a cleaner way to do this.
   // The 3 interfaces are position, velocity, and effort.
-  int interfaces_per_joint_ = 3;
+  size_t interfaces_per_joint_ = 3;
 
   std::unique_ptr<::bosdyn::client::Robot> robot_;
   ::bosdyn::client::LeaseClient* lease_client_;
